refactor(stt): extract clipboard/history/ui-sounds from stt_daemon god object (ADR-059 V10)
- clipboard.py: write_clipboard + auto_paste (public names, per P1-5)
- history.py: append_history, wav_duration_s, HISTORY_PATH alias, read_history
- ui_sounds.py: play_ui_sound
- paths.py: add STT_HISTORY_PATH constant
- stt_daemon.py: 993 → 797 lines; all extracted symbols imported from new modules
- cli.py: import from clipboard/history instead of stt_daemon internals
- tests: update patch targets to new public names (play_ui_sound, write_clipboard)
